switch delay test 0.0.5

+++ script/health.sh
@@ -23,7 +23,10 @@ do
   then
     echo "> Health check 성공"
     switch_proxy
-
+    if [ "$?" -ge 1 ]
+    then
+      exit 1
+    fi
     echo "> Nginx에 연결되지 않은 container 삭제"
     IDLE_CONTAINER=$(find_idle_profile)
     docker-compose -f docker-compose.prod.yml rm -s -v -f "$IDLE_CONTAINER"
